requirements.txt
google-genai>=1.0.0
PyMuPDF>=1.23.0
pdfplumber>=0.10.0
reportlab>=4.0.0
pytesseract>=0.3.10
Pillow>=10.0.0
pytest>=7.0.0
